I can make a square now (also use uint32 for put_pixel)

+++ include/so_long.h
@@ -6,7 +6,7 @@
 /*   By: estettle <estettle@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 15:19:22 by estettle          #+#    #+#             */
-/*   Updated: 2024/11/27 10:31:41 by estettle         ###   ########.fr       */
+/*   Updated: 2024/11/27 11:28:18 by estettle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,9 @@
 // MinilibX
 # include "../mlx_linux/mlx.h"
 
+// Stdlib
+# include <inttypes.h>
+
 // Structs
 typedef struct s_data
 {
@@ -30,6 +33,6 @@ typedef struct s_data
 }	t_data;
 
 // Window utils functions
-void	put_pixel(t_data *data, int x, int y, long color);
+void	put_pixel(t_data *data, int x, int y, uint32_t color);
 
 #endif //SO_LONG_H
